requirements.txt
playwright>=1.49.0
python-dotenv>=1.0.1
pydantic>=2.10.0
streamlit>=1.41.0
pandas>=2.2.0
openpyxl>=3.1.0
aiohttp>=3.11.0
rich>=13.9.0
